refactor(observer): split reconcile by responsibility
