+++ apps/observer/src/reconcile/reconcileResult.ts
@@ -1,0 +1,14 @@
+import type { SafeError } from "@station/contracts";
+
+export type ReconcileTiming = {
+  reason: string;
+  startedAt: string;
+  finishedAt: string;
+  durationMs: number;
+  projectsScanned: number;
+  worktreesObserved: number;
+  terminalTargetsObserved: number;
+  harnessRunsObserved: number;
+  eventsEmitted: number;
+  errors: SafeError[];
+};
